Handle false from DateTime::getLastErrors on php8.2
Also avoid implicit conversion from float 48.8 to int loses precision

+++ src/Auth/Password.php
@@ -89,7 +89,7 @@ class Password {
 		$random = self::getBytes( 16 );
 
 		$output = self::BLOWFISH_PREFIX;
-		$output .= chr( ord( '0' ) + $cost / 10 );
+		$output .= chr( ord( '0' ) + (int)( $cost / 10 ) );
 		$output .= chr( ord( '0' ) + $cost % 10 );
 		$output .= '$';
